+++ src/main/java/net/geforcemods/securitycraft/blockentities/FloorTrapBlockEntity.java
@@ -25,8 +25,8 @@ import net.minecraft.world.phys.AABB;
 public class FloorTrapBlockEntity extends DisguisableBlockEntity implements ITickingBlockEntity {
 	private IgnoreOwnerOption ignoreOwner = new IgnoreOwnerOption(true);
 	private BooleanOption disappearInstantlyInChains = new BooleanOption("disappearInstantlyInChains", true);
-	private IntOption disappearDelay = new IntOption("disappearDelay", 5, 0, 200, 1, true);
-	private IntOption reappearDelay = new IntOption("reappearDelay", 20, 5, 200, 1, true);
+	private IntOption disappearDelay = new IntOption("disappearDelay", 5, 0, 200, 1);
+	private IntOption reappearDelay = new IntOption("reappearDelay", 20, 5, 200, 1);
 	private TargetingModeOption targetingMode = new TargetingModeOption(TargetingMode.PLAYERS);
 	private boolean shouldDisappear = false, shouldReappear = false;
 	private int ticksUntilDisappearing = -1, ticksUntilReappearing = -1;