+++ src/main/java/net/geforcemods/securitycraft/blockentities/RiftStabilizerBlockEntity.java
@@ -45,8 +45,8 @@ import net.neoforged.neoforge.event.entity.EntityTeleportEvent.TeleportCommand;
 import net.neoforged.neoforge.network.PacketDistributor;
 
 public class RiftStabilizerBlockEntity extends DisguisableBlockEntity implements ITickingBlockEntity, ILockable, IToggleableEntries<TeleportationType> {
-	private final IntOption signalLength = new IntOption("signalLength", 60, 0, 400, 5, true); //20 seconds max
-	private final IntOption range = new IntOption("range", 5, 1, 15, 1, true);
+	private final IntOption signalLength = new IntOption("signalLength", 60, 0, 400, 5); //20 seconds max
+	private final IntOption range = new IntOption("range", 5, 1, 15, 1);
 	private final DisabledOption disabled = new DisabledOption(false);
 	private IgnoreOwnerOption ignoreOwner = new IgnoreOwnerOption(true);
 	private final Map<TeleportationType, Boolean> teleportationFilter = new EnumMap<>(TeleportationType.class);